cloud_function_summarizer/requirements.txt
google-cloud-storage
google-cloud-firestore
transformers
torch
PyMuPDF
